requirements.txt
crewai==0.121.0
langchain-anthropic==0.3.12
tavily-python==0.5.1
python-dotenv==1.1.0
